[Bgpcfgd] Add dependency callback logic in SRv6 Manager to handle out-of-order table processing (#794)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log skip_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it

Previously, if the SRv6 manager received a config update that added a SRV6_MY_SID entry before its corresponding LOCATOR entry added, the SRv6 manager would not process the SRV6_MY_SID entry and print an ERR syslog. However, during testing, NVIDIA team found that it is possible to have such a scenario in normal operations, especially during config reload. So, we decide to add a dependency-based caching mechanism to handle out-of-order table processing.

This is the first use case where we will dynamically add/remove the dependency, so we also define a new unsubscribe function for the directory class. The reason why we need to remove the dependency in managers_srv6 is because the on_deps_change function will only process the cache when all dependencies are satisfied. If a locator is removed and the corresponding dependency is not removed, then the on_deps_change will never get to the satisfied condition.

Fix #21826 (Issue)

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

#### How to verify it

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 201811
- [ ] 201911
- [ ] 202006
- [ ] 202012
- [ ] 202106
- [ ] 202111
- [ ] 202205
- [ ] 202211
- [ ] 202305
- [x] 202412

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [x] 202412

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)