[Arista] Update platform library submodules (#962)
#### Why I did it
Add support for QuicksilverP512 platform.

Note: additional changes from January are included in this update:
improvements for the Moby and Wolverine platforms.

This cherry-pick is specific to the 202412 branch.

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it
Updated both submodules to update the driver code.

#### How to verify it
Boot the image on the target device and validate that SONiC loads and
properly identifies the platform using `show version`.

#### Which release branch to backport (provide reason below if selected)
- [x] 202412

#### Description for the changelog
Update Arista platform driver submodules.