[202412][Mellanox] Update SAI version to SAIBuild2412.31.15.0 (#1094)
#### Why I did it
To update the Mellanox SAI version to `SAIBuild2412.31.15.0`

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

#### How to verify it
Ran regression and verified on 202412
<!--
If PR needs to be backported, then the PR must be tested against the
base branch and the earliest backport release branch and provide tested
image version on these two branches. For example, if the PR is requested
for master, 202211 and 202012, then the requester needs to provide test
results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

#### Tested branch (Please provide the tested image version)

#### Description for the changelog

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on
https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)