[action] [PR:22573] [Mellanox] Optimize thermal updater performance (#1109)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it

Fix issue https://github.com/sonic-net/sonic-buildimage/issues/22505

thermalctld CPU usage is high on NVIDIA platform. It is because there is a thread "thermal updater" who prepares thermal data for NVIDIA thermal control algorithm. I did some optimization regarding to the thread. After the optimization, the CPU usage of thermalctld changes from 20% ~ 2%. (I tested on sn5600 with 32 optical modules, the test results may be different on different platform with different number of optical modules)

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

1. Combine `get_temperature`, `get_temperature_warning_threshold`, `get_temperature_critical_threshold` into one function `get_temperature_info` to avoid duplicate function calls such as "self.is_sw_control"
2. Previously, `get_temperature`, `get_temperature_warning_threshold`, `get_temperature_critical_threshold` call `self.reinit` each time to handle module replacement. However, `self.reinit` is a expensive call according to profile result. This PR compares module SN to detect module replacement. It only calls `self.reinit` if module SN changed
3. Previously, each call to `get_temperature_warning_threshold` and `get_temperature_critical_threshold` will read all types of threshold of the module. This is not necessary. This PR only gets high/critical temperature threshold to reduce EEPROM reading effort.

#### How to verify it

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

1. manual test
2. unit test

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 201811
- [ ] 201911
- [ ] 202006
- [ ] 202012
- [ ] 202106
- [ ] 202111
- [ ] 202205
- [ ] 202211
- [ ] 202305

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)